backend/cloud-backend/inference-service/requirements.txt
keras==3.12.1
keras-hub==0.25.1
tensorflow==2.20.0
jax==0.6.2
jaxlib==0.6.2
numpy==2.2.6
pandas==2.3.3
scipy==1.15.3
librosa==0.11.0
joblib==1.5.3
scikit-learn==1.6.1

fastapi==0.115.8
uvicorn[standard]==0.34.0
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
pydantic==2.10.6
